Add logic for listing soft deleted application, deleting permanently applications and recover them

+++ src/main/java/com/tu/codeguard/service/impl/ApplicationServiceImpl.java
@@ -3,6 +3,7 @@ package com.tu.codeguard.service.impl;
 import com.tu.codeguard.dbo.ApplicationEntity;
 import com.tu.codeguard.dto.Application;
 import com.tu.codeguard.dto.ApplicationDTO;
+import com.tu.codeguard.exceptions.EntityNotFoundException;
 import com.tu.codeguard.repository.ApplicationJpaRepository;
 import com.tu.codeguard.service.ApplicationService;
 import com.tu.codeguard.utils.Mapper;
@@ -12,6 +13,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Slf4j
 @Service
@@ -26,7 +28,22 @@ public class ApplicationServiceImpl implements ApplicationService {
     @Override
     public List<ApplicationDTO> getAllApplications() {
         log.debug("Retrieving all applications");
-        List<ApplicationEntity> entities = applicationJpaRepository.findAll();
+        List<ApplicationEntity> entities = applicationJpaRepository
+                .findAll()
+                .stream()
+                .filter(entity -> !entity.isDeleted())
+                .toList();
+        return entities.stream().map(Mapper::mapApplicationToApplicationDTO).toList();
+    }
+
+    @Override
+    public List<ApplicationDTO> getAllDeletedApplications() {
+        log.debug("Retrieving all deleted applications");
+        List<ApplicationEntity> entities = applicationJpaRepository
+                .findAll()
+                .stream()
+                .filter(ApplicationEntity::isDeleted)
+                .toList();
         return entities.stream().map(Mapper::mapApplicationToApplicationDTO).toList();
     }
 
@@ -34,6 +51,7 @@ public class ApplicationServiceImpl implements ApplicationService {
     public String getApplicationDetailsById(String id) {
         ApplicationEntity entity = applicationJpaRepository.findById(id).orElse(null);
         if (entity == null) { return null; }
+        if (entity.isDeleted()) { return null; }
 
         return s3Service.downloadTxtFile(entity.getAiResultFilePath());
     }
@@ -47,7 +65,41 @@ public class ApplicationServiceImpl implements ApplicationService {
 
     @Override
     public void deleteApplicationById(String id) {
-        log.debug("Deleting application by id: {}", id);
+        log.debug("Deleting permanently application by id: {}", id);
+
+        Optional<ApplicationEntity> applicationEntity = applicationJpaRepository.findById(id);
+
+        if (applicationEntity.isEmpty()) {
+            throw new EntityNotFoundException("Entity with id " + id + " not found when trying to delete");
+        }
+
+        s3Service.delete(applicationEntity.get().getAiResultFilePath());
         applicationJpaRepository.deleteById(id);
+    }
+
+    @Override
+    public void softDeleteApplicationById(String id) {
+        log.debug("Soft deleting application by id: {}", id);
+        Optional<ApplicationEntity> applicationEntity = applicationJpaRepository.findById(id);
+
+        if (applicationEntity.isEmpty()) {
+            throw new EntityNotFoundException("Entity with id " + id + " not found when trying to softDelete");
+        }
+
+        applicationEntity.get().setDeleted(true);
+        applicationJpaRepository.save(applicationEntity.get());
+    }
+
+    @Override
+    public void recoverDeletedApplicationById(String id) {
+        log.debug("Recover deleted application by id: {}", id);
+        Optional<ApplicationEntity> applicationEntity = applicationJpaRepository.findById(id);
+
+        if (applicationEntity.isEmpty()) {
+            throw new EntityNotFoundException("Entity with id " + id + " not found when trying to recover");
+        }
+
+        applicationEntity.get().setDeleted(false);
+        applicationJpaRepository.save(applicationEntity.get());
     }
 }
